Update test, but it's broken
